+++ bridge_front/src/App.js
@@ -5,12 +5,12 @@ import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css'
 import Home from './components/home';
 import "./App.css"
+import Exercise from './exercise/Exercise';
 
 function App() {
   return (
     <div className="App">
-      <NavBar/>
-      <Home/>
+      <Exercise />
     </div>
   );
 }
