+++ bridge_front/src/exercise/ExerciseContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react"; 
+
+export const ExerciseContext  = createContext(null);
